ref: simplify VueScreenerTableView
